+++ packages/hydrogen/CHANGELOG.md
@@ -51,6 +51,7 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 - fix: Clear browser fetch cache by @wizardlyhel in [#591](https://github.com/Shopify/hydrogen/pull/591)
 - refactor: use featureImage instead of images(first:1) on product query
 - fix: target future release to use '2022-01' API Version
+- fix: `ShopPayButton` supports quantities greater than 1. Also fixed issues with IDs in Storefront API version 2022-01
 
 ## 0.10.1 - 2022-01-26